ubuntu 18.04 gcc7.2 complie 完成

+++ include/libBCH/libBCH.hpp
@@ -95,7 +95,7 @@ namespace jub {
 			for (auto output : outputs)
 			{
 				//amount
-				unsinged_trans << output.amount;
+				unsinged_trans << (uint64_t)output.amount;
 				//script_pub
 				uchar_vector script_pub;
 				if (JUBR_OK != buildScriptPubFromAddress(output.address, script_pub))
